SmartSurgSim/backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
channels>=4.0
djangorestframework-simplejwt>=5.3
pytest>=8.0
pytest-django>=4.8
